restructure collecting and reading of files
