+++ src/grammar/sakura_parser.y
@@ -43,7 +43,6 @@
 #include <vector>
 #include <libKitsunemimiCommon/common_items/data_items.h>
 #include <items/sakura_items.h>
-#include <parsing/sakura_parsing.h>
 
 using Kitsunemimi::DataItem;
 using Kitsunemimi::DataArray;
@@ -664,7 +663,6 @@ subtree_fork:
         $$ = new SubtreeItem();
         $$->nameOrPath = $3;
         $$->values = *$5;
-        driver.m_sakuraParsing->addFileToQueue($3);
         delete $5;
     }
