+++ tests/functional_tests/functional_tests.pro
@@ -40,3 +40,8 @@ SOURCES += \
 HEADERS += \
     test_blossom.h \
     interface_test.h
+
+
+unix {
+    INCLUDEPATH += /usr/lib/gcc/x86_64-linux-gnu/9/include
+}
